+++ src/components/Login/LoginFormContainer.tsx
@@ -4,13 +4,13 @@ import {
   StyledLoginComponentDiv,
   StyledLoginKakoLogo,
 } from '@src/components/Login/StyledLoginFormContainer';
-import { useLoginForm } from './useLoginForm';
+import { useLoginForm } from '../../hooks/useLoginForm';
 import { apiClient } from '@src/api/FetchData';
 import { BASIC_ENDPOINT } from '@src/assets/endpoints';
 import { useNavigate } from 'react-router-dom';
 import { URLS } from '@src/assets/urls';
 import { SESSION_KEY_NAME } from '@src/assets/sessionId';
-import type { HttpsFailedResponseTypes } from './LoginFetchDataType';
+import type { HttpsFailedResponseTypes } from '../../types/LoginFetchDataType';
 import { toast, ToastContainer } from 'react-toastify';
 
 const LoginForm = () => {
